+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   * [Fixed] ES6 import dependency on jquery for jquery.fancytree.ui-deps.js
   * [Fixed] #863 setActive() sometimes does not scroll node into view
   * [Fixed] #877 postProcess may now also return the object form `{..., children: []}`
+  * [Fixed] #884 ReferenceError: jQuery is not defined at _simpleDeepMerge
     (before, only an array `[...]` was handled correctly).
   * [Fixed] autoScroll, node.scrollIntoView(), and .makeVisible() now work for tables as well.
   * [Changed] #879 Rename ext-debug to ext-logger